Created frontend react application including the authentication forms

+++ backend/controllers/goalController.js
@@ -1,10 +1,12 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("../models/goalModel");
+const User = require("../models/userModel");
+
 //@desc Get goals
 //@route GET /api/goals
 //@access private
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user.id });
   res.status(200).json(goals);
 });
 
@@ -18,6 +20,7 @@ const postGoal = asyncHandler(async (req, res) => {
   }
 
   const goal = await Goal.create({
+    user: req.user.id,
     text: req.body.text,
   });
   res.status(200).json(goal);
@@ -38,6 +41,20 @@ const updateGoal = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("Please provide a text value");
   }
+
+  const user = await User.findById(req.user.id);
+  //check for user
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  //make sure the logged in user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
+
   const updateGoal = await Goal.findByIdAndUpdate(id, req.body, {
     new: true,
   });
@@ -54,6 +71,21 @@ const deleteGoal = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("Goal does not exist");
   }
+  const user = await User.findById(req.user.id);
+  //check for user
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  //make sure the logged in user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
+
+  //or
+  // await goal.remove();
   const deletedGoal = await Goal.findByIdAndRemove(id);
   res.status(200).json(goal);
 });
